+++ adapters/simulator/src/main/kotlin/com/amazonaws/sfc/simulator/simulations/List.kt
@@ -3,8 +3,8 @@
 
 package com.amazonaws.sfc.simulator.simulations
 
-import com.amazonaws.sfc.simulator.config.SimulationConfiguration.Companion.CONFIG_DATA_TYPE
-import com.amazonaws.sfc.simulator.config.SimulationConfiguration.Companion.CONFIG_ITEMS
+import com.amazonaws.sfc.simulator.config.SimulatorConfiguration.Companion.CONFIG_DATA_TYPE
+import com.amazonaws.sfc.simulator.config.SimulatorConfiguration.Companion.CONFIG_ITEMS
 import com.amazonaws.sfc.simulator.simulations.SimulationDeserializer.Companion.createSimulationReader
 import com.amazonaws.sfc.simulator.simulations.SimulationHelper.asAnyType
 import com.google.gson.JsonObject