+++ core/sfc-core/src/main/kotlin/com/amazonaws/sfc/client/AwsServiceClientHelper.kt
@@ -21,6 +21,7 @@ import software.amazon.awssdk.awscore.internal.AwsErrorCode
 import software.amazon.awssdk.core.SdkClient
 import software.amazon.awssdk.core.client.config.ClientOverrideConfiguration
 import software.amazon.awssdk.core.client.config.SdkAdvancedClientOption
+import java.net.URI
 
 
 // Internal exception class used to indicate that failed execution of a code block making the AWS
@@ -93,6 +94,10 @@ abstract class AwsServiceClientHelper(
         val region = awsService?.region
         if (region != null) {
             builder.region(region)
+        }
+
+        if (awsService?.endpoint != null) {
+            builder.endpointOverride(URI(awsService?.endpoint?:""))
         }
 
         val clientCredentialsProvider = getCredentialsProvider()