+++ targets/aws-sns-target/src/main/kotlin/com/amazonaws/sfc/awssns/config/AwsSnsTargetConfiguration.kt
@@ -8,6 +8,7 @@ package com.amazonaws.sfc.awssns.config
 import com.amazonaws.sfc.awssns.config.AwsSnsWriterConfiguration.Companion.AWS_SNS
 import com.amazonaws.sfc.config.AwsServiceConfig
 import com.amazonaws.sfc.config.BaseConfiguration.Companion.CONFIG_BATCH_SIZE
+import com.amazonaws.sfc.config.BaseConfiguration.Companion.CONFIG_ENDPOINT
 import com.amazonaws.sfc.config.BaseConfiguration.Companion.CONFIG_INTERVAL
 import com.amazonaws.sfc.config.BaseConfiguration.Companion.CONFIG_REGION
 import com.amazonaws.sfc.config.ConfigurationClass
@@ -72,6 +73,11 @@ class AwsSnsTargetConfiguration : AwsServiceConfig, TargetConfiguration() {
      */
     override val region: Region?
         get() = if (_region.isNullOrEmpty()) null else Region.of(_region!!.lowercase())
+
+    @SerializedName(CONFIG_ENDPOINT)
+    var _endPoint : String? = null
+    override val endpoint : String?
+        get() = _endPoint
 
     @SerializedName(CONFIG_INTERVAL)
     private var _interval: Int? = null
@@ -170,6 +176,7 @@ class AwsSnsTargetConfiguration : AwsServiceConfig, TargetConfiguration() {
                    messageGroupId: String? = default._messageGroupId,
                    serialAsMessageDeduplicationId: Boolean = default._serialAsMessageDeduplicationId,
                    region: String? = default._region,
+                   endPoint: String? = default._endPoint,
                    interval: Int? = default._interval,
                    batchSize: Int = default._batchSize,
                    description: String = default._description,
@@ -193,6 +200,7 @@ class AwsSnsTargetConfiguration : AwsServiceConfig, TargetConfiguration() {
                 _messageGroupId = messageGroupId
                 _serialAsMessageDeduplicationId = serialAsMessageDeduplicationId
                 _region = region
+                _endPoint = endPoint
                 _interval = interval
                 _batchSize = batchSize
             }