+++ targets/aws-timestream-target/src/main/kotlin/com/amazonaws/sfc/awstimestream/config/AwsTimestreamTargetConfiguration.kt
@@ -8,6 +8,7 @@ package com.amazonaws.sfc.awstimestream.config
 import com.amazonaws.sfc.awstimestream.config.AwsTimestreamWriterConfiguration.Companion.AWS_TIMESTREAM
 import com.amazonaws.sfc.config.*
 import com.amazonaws.sfc.config.BaseConfiguration.Companion.CONFIG_BATCH_SIZE
+import com.amazonaws.sfc.config.BaseConfiguration.Companion.CONFIG_ENDPOINT
 import com.amazonaws.sfc.config.BaseConfiguration.Companion.CONFIG_INTERVAL
 import com.amazonaws.sfc.config.BaseConfiguration.Companion.CONFIG_REGION
 import com.amazonaws.sfc.metrics.MetricsSourceConfiguration
@@ -44,6 +45,11 @@ class AwsTimestreamTargetConfiguration : AwsServiceConfig, TargetConfiguration()
      */
     override val region: Region?
         get() = if (_region.isNullOrEmpty()) null else Region.of(_region!!.lowercase())
+
+    @SerializedName(CONFIG_ENDPOINT)
+    var _endPoint : String? = null
+    override val endpoint : String?
+        get() = _endPoint
 
     @SerializedName(CONFIG_BATCH_SIZE)
     private var _batchSize: Int = 10
@@ -143,6 +149,7 @@ class AwsTimestreamTargetConfiguration : AwsServiceConfig, TargetConfiguration()
         fun create(tableName: String? = default._tableName,
                    database: String? = default._database,
                    region: String? = default._region,
+                   endpoint : String? = default._endPoint,
                    batchSize: Int = default._batchSize,
                    interval: Int? = default._interval,
                    records: List<AwsTimestreamRecordConfiguration> = emptyList(),
@@ -165,6 +172,7 @@ class AwsTimestreamTargetConfiguration : AwsServiceConfig, TargetConfiguration()
                 _tableName = tableName
                 _database = database
                 _region = region
+                _endPoint = endpoint
                 _batchSize = batchSize
                 _interval = interval
                 _records = records